SPEAKEASY-35: Fix modules not being enabled after startup or reinstall of speakeasy
* Change state id to use bundle last modified date

+++ src/main/java/com/atlassian/labs/speakeasy/DescriptorGenerator.java
@@ -10,5 +10,5 @@ import java.util.List;
  */
 public interface DescriptorGenerator<D extends ModuleDescriptor>
 {
-    Iterable<D> getDescriptorsToExposeForUsers(List<String> users, int state);
+    Iterable<D> getDescriptorsToExposeForUsers(List<String> users, long state);
 }
